Change data types from double to decimal in view models
Updated the data types of the `GrandTotal` property in the
`OrderIndexViewModel` class and the `TotalItemPrice` property
in the `OrderItemViewModel` class from `double` to `decimal`.
This change enhances precision in financial calculations,
as `decimal` is more suitable for representing monetary values.

+++ ST10361554-CLDV6212-POE-Part-3-WebApp/ViewModels/OrderIndexViewModel.cs
@@ -12,6 +12,6 @@
         public required string OrderStatus { get; set; }
 
         // Total amount for the order
-        public double GrandTotal { get; set; }
+        public decimal GrandTotal { get; set; }
     }
 }
